crewai_example/requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
python-dotenv
